--- src/main/java/site/dittotrip/ditto_trip/review/service/ReviewRemoveService.java
@@ -1,29 +1,0 @@
-package site.dittotrip.ditto_trip.review.service;
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-import site.dittotrip.ditto_trip.review.domain.Review;
-import site.dittotrip.ditto_trip.review.repository.ReviewRepository;
-import site.dittotrip.ditto_trip.user.domain.User;
-
-import java.util.NoSuchElementException;
-
-@Service
-@Transactional(readOnly = false)
-@RequiredArgsConstructor
-public class ReviewRemoveService {
-
-    private ReviewRepository reviewRepository;
-
-    public void removeReview(Long reviewId, User user) {
-        Review review = reviewRepository.findById(reviewId).orElseThrow(NoSuchElementException::new);
-
-        if (!review.getUser().equals(user)) {
-            // throw new NoAuthorityException();
-        }
-
-        reviewRepository.delete(review);
-    }
-
-}
